+++ packages/auto_mappr/lib/src/builder/methods/mapping_method_builder.dart
@@ -1,5 +1,6 @@
 import 'package:auto_mappr/src/builder/map_model_body_method_builder.dart';
 import 'package:auto_mappr/src/builder/methods/method_builder_base.dart';
+import 'package:auto_mappr/src/extensions/expression_extension.dart';
 import 'package:auto_mappr/src/helpers/emitter_helper.dart';
 import 'package:auto_mappr/src/models/type_mapping.dart';
 import 'package:code_builder/code_builder.dart';
@@ -9,7 +10,7 @@ class MappingMethodBuilder extends MethodBuilderBase {
   final bool nullable;
   final void Function(TypeMapping? mapping)? usedNullableMethodCallback;
 
-  MappingMethodBuilder(
+  const MappingMethodBuilder(
     super.config, {
     required this.mapping,
     this.nullable = false,
@@ -18,49 +19,35 @@ class MappingMethodBuilder extends MethodBuilderBase {
 
   @override
   Method buildMethod() {
-    // TODO(later): rework mapping to this builder
+    var returnType = EmitterHelper.current.typeRefer(type: mapping.target);
+
+    if (nullable) {
+      returnType = returnType.nullabled();
+    }
 
     return Method(
       (b) => b
-        ..name = mapping.mappingMethodName(config: config)
+        ..name =
+            nullable ? mapping.nullableMappingMethodName(config: config) : mapping.mappingMethodName(config: config)
         ..requiredParameters.addAll([
           Parameter(
             (p) => p
               ..name = 'input'
-              ..type = EmitterHelper.current.typeRefer(type: mapping.source),
+              ..type = EmitterHelper.current.typeRefer(type: mapping.source).nullabled(),
           ),
         ])
-        ..returns = EmitterHelper.current.typeRefer(type: mapping.target)
-        ..body = MapModelBodyMethodBuilder(
-          mapping: mapping,
-          mapperConfig: config,
-          usedNullableMethodCallback: usedNullableMethodCallback,
-        ).build(),
+        ..returns = returnType
+        ..body = buildBody(),
     );
-
-    // nullable
-    // return Method(
-    //   (b) => b
-    //     ..name = mapping.nullableMappingMethodName
-    //     ..requiredParameters.addAll([
-    //       Parameter(
-    //         (p) => p
-    //           ..name = 'input'
-    //           ..type = refer('${mapping.source.getDisplayString(withNullability: false)}?'),
-    //       )
-    //     ])
-    //     ..returns = refer('${mapping.target.getDisplayString(withNullability: true)}?')
-    //     ..body = MapModelBodyMethodBuilder(
-    //       mapping: mapping,
-    //       mapperConfig: config,
-    //       nullable: true,
-    //     ).build(),
-    // );
   }
 
   @override
   Code buildBody() {
-    // TODO(later): rework
-    return const Code('T');
+    return MapModelBodyMethodBuilder(
+      mapping: mapping,
+      mapperConfig: config,
+      usedNullableMethodCallback: usedNullableMethodCallback,
+      nullable: nullable,
+    ).build();
   }
 }