+++ packages/auto_mappr/lib/src/extensions/reference_extension.dart
@@ -1,6 +1,0 @@
-import 'package:auto_mappr/src/helpers/emitter_helper.dart';
-import 'package:code_builder/code_builder.dart';
-
-extension ReferenceExtension on Reference {
-  Reference get nullabled => refer('${accept(EmitterHelper.current.emitter)}?');
-}