+++ packages/auto_mappr/test/builder/convert_method_builder_test.dart
@@ -2,7 +2,7 @@ import 'package:auto_mappr/src/builder/methods/convert_method_builder.dart';
 import 'package:auto_mappr/src/builder/methods/private_convert_method_builder.dart';
 import 'package:auto_mappr/src/builder/methods/try_convert_method_builder.dart';
 import 'package:auto_mappr/src/builder/methods/type_of_method_builder.dart';
-import 'package:auto_mappr/src/extensions/reference_extension.dart';
+import 'package:auto_mappr/src/extensions/expression_extension.dart';
 import 'package:auto_mappr/src/helpers/run_zoned_auto_mappr.dart';
 import 'package:auto_mappr/src/models/auto_mappr_config.dart';
 import 'package:auto_mappr/src/models/auto_mappr_options.dart';
@@ -13,8 +13,8 @@ import 'package:test/test.dart';
 void main() {
   test('buildConvertMethod has the correct interface', () {
     runZonedAutoMappr(() {
-      final result = ConvertMethodBuilder(
-        const AutoMapprConfig(
+      final result = const ConvertMethodBuilder(
+        AutoMapprConfig(
           mappers: [],
           availableMappingsMacroId: 'test',
           mapprOptions: AutoMapprOptions(ignoreNullableSourceField: false),
@@ -32,7 +32,7 @@ void main() {
           Parameter(
             (p) => p
               ..name = 'model'
-              ..type = result.types.firstOrNull?.nullabled,
+              ..type = result.types.firstOrNull?.nullabled(),
           ),
         ]),
       );
@@ -41,8 +41,8 @@ void main() {
 
   test('buildTryConvertMethod has the correct interface', () {
     runZonedAutoMappr(() {
-      final result = TryConvertMethodBuilder(
-        const AutoMapprConfig(
+      final result = const TryConvertMethodBuilder(
+        AutoMapprConfig(
           mappers: [],
           availableMappingsMacroId: 'test',
           mapprOptions: AutoMapprOptions(ignoreNullableSourceField: false),
@@ -52,7 +52,7 @@ void main() {
       expect(result, isA<Method>());
       expect(result.name, equals('tryConvert'));
       expect(result.types.length, equals(2));
-      expect(result.returns, equals(result.types.elementAtOrNull(1)?.nullabled));
+      expect(result.returns, equals(result.types.elementAtOrNull(1)?.nullabled()));
       expect(result.optionalParameters, isEmpty);
       expect(
         result.requiredParameters,
@@ -60,7 +60,7 @@ void main() {
           Parameter(
             (p) => p
               ..name = 'model'
-              ..type = result.types.firstOrNull?.nullabled,
+              ..type = result.types.firstOrNull?.nullabled(),
           ),
         ]),
       );
@@ -69,8 +69,8 @@ void main() {
 
   test('buildInternalConvertMethod has the correct interface', () {
     runZonedAutoMappr(() {
-      final result = PrivateConvertMethodBuilder(
-        const AutoMapprConfig(
+      final result = const PrivateConvertMethodBuilder(
+        AutoMapprConfig(
           mappers: [],
           availableMappingsMacroId: 'test',
           mapprOptions: AutoMapprOptions(ignoreNullableSourceField: false),
@@ -80,7 +80,7 @@ void main() {
       expect(result, isA<Method>());
       expect(result.name, equals('_convert'));
       expect(result.types.length, equals(2));
-      expect(result.returns, equals(result.types.elementAtOrNull(1)?.nullabled));
+      expect(result.returns, equals(result.types.elementAtOrNull(1)?.nullabled()));
       expect(result.optionalParameters.isEmpty, isFalse);
       expect(
         result.requiredParameters,
@@ -88,7 +88,7 @@ void main() {
           Parameter(
             (p) => p
               ..name = 'model'
-              ..type = result.types.firstOrNull?.nullabled,
+              ..type = result.types.firstOrNull?.nullabled(),
           ),
         ]),
       );
@@ -97,8 +97,8 @@ void main() {
 
   test('buildTypeOfHelperMethod has the correct interface', () {
     runZonedAutoMappr(() {
-      final result = TypeOfMethodBuilder(
-        const AutoMapprConfig(
+      final result = const TypeOfMethodBuilder(
+        AutoMapprConfig(
           mappers: [],
           availableMappingsMacroId: 'test',
           mapprOptions: AutoMapprOptions(ignoreNullableSourceField: false),